+++ pallets/briefs/Cargo.toml
@@ -6,7 +6,7 @@ version = '0.1.0'
 license = 'Apache 2.0'
 homepage = 'https://github.com/ImbueNetwork/imbue'
 repository = "https://github.com/ImbueNetwork/imbue"
-edition = '2021'
+edition = '2018'
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
@@ -27,11 +27,12 @@ sp-api = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.37", default-features = false }
 sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.37", default-features = false }
 # Code coverage wrapper.
-cargo-llvm-cov = "0.5.11"
+#cargo-llvm-cov = "0.5.11"
 # local dependancies
 common-traits = { path = "../../libs/common-traits", default-features = false }
 common-types = { path = "../../libs/common-types", default-features = false }
 proposals = {path= "../proposals", default-features = false } 
+pallet-timestamp = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.37", default-features = false }
 
 
 [dev-dependencies]
@@ -65,6 +66,7 @@ std = [
 	"pallet-balances/std",
 	"proposals/std",
 	"pallet-identity/std",
+	"pallet-timestamp/std",
 	]
 runtime-benchmarks = ["frame-benchmarking/runtime-benchmarks"]
 try-runtime = ["frame-support/try-runtime"]
